split header into separate widget files for search dropdown, drawer and notification badge
