+++ thryft/lib/widgets/header.dart
@@ -6,6 +6,8 @@ import 'package:thryft/providers/notification_provider.dart';
 import 'package:thryft/providers/search_provider.dart';
 import 'package:thryft/utils/responsive.dart';
 import 'package:thryft/widgets/filter_system.dart';
+import 'package:thryft/widgets/notification_badge.dart';
+import 'package:thryft/widgets/search_dropdown.dart';
 
 class Header extends StatelessWidget {
   const Header({super.key});
@@ -64,7 +66,7 @@ class _DesktopHeaderState extends State<_DesktopHeader> {
   void _showOverlay() {
     _removeOverlay();
     _overlayEntry = OverlayEntry(
-      builder: (_) => _SearchDropdown(
+      builder: (_) => SearchDropdown(
         layerLink: _layerLink,
         controller: _searchController,
         dropdownWidth: 500,
@@ -188,7 +190,7 @@ class _DesktopHeaderState extends State<_DesktopHeader> {
                           final session =
                               Supabase.instance.client.auth.currentSession;
                           if (session == null) return const SizedBox.shrink();
-                          return _NotificationBadge(
+                          return NotificationBadge(
                             count: notifProvider.unreadCount,
                             onPressed: () => context.push('/notifications'),
                           );
@@ -365,7 +367,7 @@ class _MobileHeaderState extends State<_MobileHeader> {
     _removeOverlay();
     final screenWidth = MediaQuery.of(context).size.width;
     _overlayEntry = OverlayEntry(
-      builder: (_) => _SearchDropdown(
+      builder: (_) => SearchDropdown(
         layerLink: _layerLink,
         controller: _searchController,
         dropdownWidth: screenWidth - 16,
@@ -451,7 +453,7 @@ class _MobileHeaderState extends State<_MobileHeader> {
             builder: (context, notifProvider, _) {
               final session = Supabase.instance.client.auth.currentSession;
               if (session == null) return const SizedBox.shrink();
-              return _NotificationBadge(
+              return NotificationBadge(
                 count: notifProvider.unreadCount,
                 onPressed: () => context.push('/notifications'),
               );
@@ -486,391 +488,3 @@ class _MobileHeaderState extends State<_MobileHeader> {
   }
 }
 
-// ─── Search dropdown overlay ─────────────────────────────────────────────────
-
-class _SearchDropdown extends StatelessWidget {
-  final LayerLink layerLink;
-  final TextEditingController controller;
-  final double dropdownWidth;
-  final VoidCallback onDismiss;
-  final VoidCallback onNavigate;
-
-  const _SearchDropdown({
-    required this.layerLink,
-    required this.controller,
-    required this.dropdownWidth,
-    required this.onDismiss,
-    required this.onNavigate,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        // Full-screen barrier: tap outside dismisses
-        Positioned.fill(
-          child: GestureDetector(
-            behavior: HitTestBehavior.translucent,
-            onTap: onDismiss,
-          ),
-        ),
-        CompositedTransformFollower(
-          link: layerLink,
-          targetAnchor: Alignment.bottomLeft,
-          followerAnchor: Alignment.topLeft,
-          showWhenUnlinked: false,
-          child: GestureDetector(
-            // Absorb taps so they don't hit the barrier above
-            onTap: () {},
-            child: Material(
-              elevation: 8,
-              borderRadius: BorderRadius.circular(8),
-              clipBehavior: Clip.antiAlias,
-              child: SizedBox(
-                width: dropdownWidth,
-                child: ConstrainedBox(
-                  constraints: const BoxConstraints(maxHeight: 380),
-                  child: Consumer<SearchProvider>(
-                    builder: (context, provider, _) {
-                      if (provider.query.isEmpty) {
-                        return _buildIdle(context, provider);
-                      }
-                      if (provider.isLoading) {
-                        return const Padding(
-                          padding: EdgeInsets.all(24),
-                          child: Center(
-                            child: SizedBox(
-                              width: 24,
-                              height: 24,
-                              child: CircularProgressIndicator(strokeWidth: 2),
-                            ),
-                          ),
-                        );
-                      }
-                      if (provider.results.isEmpty) {
-                        return Padding(
-                          padding: const EdgeInsets.all(16),
-                          child: Text(
-                            'No results for "${provider.query}"',
-                            style: TextStyle(
-                              color: Colors.grey[600],
-                              fontSize: 14,
-                            ),
-                          ),
-                        );
-                      }
-                      return _buildResults(context, provider);
-                    },
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildIdle(BuildContext context, SearchProvider provider) {
-    if (provider.recentSearches.isEmpty) {
-      return Padding(
-        padding: const EdgeInsets.all(16),
-        child: Text(
-          'Start typing to search...',
-          style: TextStyle(color: Colors.grey[500], fontSize: 14),
-        ),
-      );
-    }
-    return ListView(
-      shrinkWrap: true,
-      padding: const EdgeInsets.symmetric(vertical: 8),
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 8, 8, 4),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              const Text(
-                'Recent searches',
-                style: TextStyle(fontWeight: FontWeight.w600, fontSize: 13),
-              ),
-              TextButton(
-                onPressed: provider.clearRecentSearches,
-                style: TextButton.styleFrom(
-                  padding: EdgeInsets.zero,
-                  minimumSize: Size.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                child: const Text('Clear all', style: TextStyle(fontSize: 12)),
-              ),
-            ],
-          ),
-        ),
-        ...provider.recentSearches
-            .take(5)
-            .map(
-              (q) => ListTile(
-                dense: true,
-                leading: const Icon(
-                  Icons.history,
-                  size: 18,
-                  color: Colors.grey,
-                ),
-                title: Text(q, style: const TextStyle(fontSize: 14)),
-                trailing: IconButton(
-                  icon: const Icon(Icons.close, size: 16, color: Colors.grey),
-                  onPressed: () => provider.removeRecentSearch(q),
-                ),
-                onTap: () {
-                  controller.text = q;
-                  provider.onQueryChanged(q);
-                },
-              ),
-            ),
-      ],
-    );
-  }
-
-  Widget _buildResults(BuildContext context, SearchProvider provider) {
-    final shown = provider.results.take(5).toList();
-    return ListView(
-      shrinkWrap: true,
-      padding: const EdgeInsets.symmetric(vertical: 8),
-      children: [
-        ...shown.map(
-          (product) => ListTile(
-            dense: true,
-            leading: SizedBox(
-              width: 36,
-              height: 36,
-              child: product.imageUrl != null
-                  ? ClipRRect(
-                      borderRadius: BorderRadius.circular(4),
-                      child: Image.network(
-                        product.imageUrl!,
-                        fit: BoxFit.cover,
-                        errorBuilder: (_, __, ___) => const Icon(
-                          Icons.image_not_supported,
-                          size: 24,
-                          color: Colors.grey,
-                        ),
-                      ),
-                    )
-                  : const Icon(
-                      Icons.image_not_supported,
-                      size: 24,
-                      color: Colors.grey,
-                    ),
-            ),
-            title: Text(
-              product.name,
-              style: const TextStyle(fontSize: 14),
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-            subtitle: Text(
-              '£${product.price.toStringAsFixed(2)} · ${product.brand}',
-              style: const TextStyle(fontSize: 12),
-            ),
-            onTap: () {
-              onDismiss();
-              context.push(
-                '/product/${product.id}',
-                extra: product.toRouteExtra(),
-              );
-            },
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-// ─── Mobile Drawer ───────────────────────────────────────────────────────────
-
-class AppDrawer extends StatelessWidget {
-  const AppDrawer({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Drawer(
-      child: ListView(
-        padding: EdgeInsets.zero,
-        children: [
-          DrawerHeader(
-            decoration: const BoxDecoration(
-              color: Color.fromARGB(255, 71, 164, 245),
-            ),
-            child: GestureDetector(
-              behavior: HitTestBehavior.opaque,
-              onTap: () {
-                Navigator.pop(context);
-                context.go('/');
-              },
-              child: Image.asset(
-                'assets/images/thyrft_logo.png',
-                fit: BoxFit.contain,
-                alignment: Alignment.centerLeft,
-              ),
-            ),
-          ),
-          ListTile(
-            leading: const Icon(Icons.home_outlined),
-            title: const Text('Home'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.checkroom_outlined),
-            title: const Text('Shirts'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/category/Shirt');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.dry_cleaning_outlined),
-            title: const Text('Trousers'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/category/Trousers');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.beach_access_outlined),
-            title: const Text('Shorts'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/category/Shorts');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.woman),
-            title: const Text('Dresses'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/category/Dresses');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.directions_walk_outlined),
-            title: const Text('Shoes'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/category/Shoes');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.watch_outlined),
-            title: const Text('Accessories'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/category/Accessories');
-            },
-          ),
-          const Divider(),
-          ListTile(
-            leading: const Icon(Icons.info_outline),
-            title: const Text('About Us'),
-            onTap: () {
-              Navigator.pop(context);
-              context.go('/about');
-            },
-          ),
-          const Divider(),
-          ListTile(
-            leading: const Icon(Icons.favorite_outline),
-            title: const Text('Wishlist'),
-            onTap: () {
-              Navigator.pop(context);
-              context.push('/wishlist');
-            },
-          ),
-          ListTile(
-            leading: const Icon(Icons.person_outline),
-            title: const Text('Account'),
-            onTap: () {
-              Navigator.pop(context);
-              final session = Supabase.instance.client.auth.currentSession;
-              if (session != null) {
-                context.push('/account');
-              } else {
-                context.push('/auth');
-              }
-            },
-          ),
-          const Divider(),
-          ListTile(
-            leading: const Icon(Icons.sell_outlined),
-            title: const Text('Sell now'),
-            onTap: () {
-              Navigator.pop(context);
-              final session = Supabase.instance.client.auth.currentSession;
-              if (session != null) {
-                context.push('/create-listing');
-              } else {
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text(
-                      'You need to be logged in to create a listing',
-                    ),
-                    duration: Duration(seconds: 3),
-                  ),
-                );
-                context.push('/auth');
-              }
-            },
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// ─── Notification Badge ───────────────────────────────────────────────────────
-
-class _NotificationBadge extends StatelessWidget {
-  final int count;
-  final VoidCallback onPressed;
-
-  const _NotificationBadge({required this.count, required this.onPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      clipBehavior: Clip.none,
-      children: [
-        IconButton(
-          icon: const Icon(Icons.notifications_outlined, color: Colors.white),
-          onPressed: onPressed,
-        ),
-        if (count > 0)
-          Positioned(
-            right: 6,
-            top: 6,
-            child: IgnorePointer(
-              child: Container(
-                padding: const EdgeInsets.all(2),
-                decoration: const BoxDecoration(
-                  color: Colors.red,
-                  shape: BoxShape.circle,
-                ),
-                constraints: const BoxConstraints(minWidth: 16, minHeight: 16),
-                child: Text(
-                  count > 99 ? '99+' : count.toString(),
-                  style: const TextStyle(
-                    color: Colors.white,
-                    fontSize: 10,
-                    fontWeight: FontWeight.bold,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-}
